Avoid static initialization order fiasco
GUIDs in wincodec.h are defined out of line, so an initializer for a static
variable can read them before they have been initialized themselves. Taking a
reference rather than copying the value bypasses this issue.

This might turn out not to be a problem in most cases, but is still undefined
behavior. Anecdotally, removing it fixed a startup crash I got on incremental
builds (so there might be some weird interaction with the linker). So it is
probably worth fixing anyway.